# internal/handler/http/auth_handler.go
package http

import (
	"encoding/json"
	"net/http"

	"github.com/crabrus/the-workshop/internal/service/auth"

	"github.com/go-chi/chi/v5"
)

// AuthHandler handles HTTP requests for authentication
type AuthHandler struct {
	authService auth.AuthService
}

// NewAuthHandler creates a new auth handler
func NewAuthHandler(service auth.AuthService) *AuthHandler {
	return &AuthHandler{
		authService: service,
	}
}

// RegisterRoutes registers all auth routes
func (h *AuthHandler) RegisterRoutes(r chi.Router) {
	r.Route("/auth", func(r chi.Router) {
		r.Post("/register", h.Register)
		r.Post("/login", h.Login)
		r.Post("/refresh", h.RefreshToken)
		r.Post("/logout", h.Logout)
	})
}

// Register godoc
// @Summary Register a new user
// @Description Create a new user account
// @Tags auth
// @Accept json
// @Produce json
// @Param request body auth.RegisterRequest true "Registration data"
// @Success 201 {object} auth.TokenResponse
// @Failure 400 {object} ErrorResponse
// @Failure 409 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /api/v1/auth/register [post]
func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req auth.RegisterRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	// Call service
	response, err := h.authService.Register(r.Context(), req)
	if err != nil {
		handleAuthError(w, err)
		return
	}

	respondJSON(w, http.StatusCreated, response)
}

// Login godoc
// @Summary Login user
// @Description Authenticate user and return tokens
// @Tags auth
// @Accept json
// @Produce json
// @Param request body auth.LoginRequest true "Login credentials"
// @Success 200 {object} auth.TokenResponse
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /api/v1/auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req auth.LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	// Call service
	response, err := h.authService.Login(r.Context(), req)
	if err != nil {
		handleAuthError(w, err)
		return
	}

	respondJSON(w, http.StatusOK, response)
}

// RefreshToken godoc
// @Summary Refresh access token
// @Description Get new access token using refresh token
// @Tags auth
// @Accept json
// @Produce json
// @Param request body RefreshTokenRequest true "Refresh token"
// @Success 200 {object} auth.TokenResponse
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /api/v1/auth/refresh [post]
func (h *AuthHandler) RefreshToken(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req RefreshTokenRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	if req.RefreshToken == "" {
		respondError(w, http.StatusBadRequest, "Refresh token is required")
		return
	}

	// Call service
	response, err := h.authService.RefreshToken(r.Context(), req.RefreshToken)
	if err != nil {
		handleAuthError(w, err)
		return
	}

	respondJSON(w, http.StatusOK, response)
}

// Logout godoc
// @Summary Logout user
// @Description Logout user (client should delete tokens)
// @Tags auth
// @Produce json
// @Success 200 {object} SuccessResponse "Logout successful"
// @Router /api/v1/auth/logout [post]
func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	// In JWT-based auth, logout is typically handled client-side
	// by deleting the tokens. This endpoint exists for consistency
	// and could be used to blacklist tokens in a production system.
	respondJSON(w, http.StatusOK, SuccessResponse{Message: "Logged out successfully"})
}

// RefreshTokenRequest represents refresh token request
type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token" validate:"required"`
}

// handleAuthError maps auth service errors to HTTP status codes
func handleAuthError(w http.ResponseWriter, err error) {
	switch err {
	case auth.ErrEmailRequired,
		auth.ErrPasswordRequired,
		auth.ErrWeakPassword:
		respondError(w, http.StatusBadRequest, err.Error())
	case auth.ErrUserAlreadyExists:
		respondError(w, http.StatusConflict, err.Error())
	case auth.ErrInvalidCredentials:
		respondError(w, http.StatusUnauthorized, err.Error())
	case auth.ErrInvalidToken,
		auth.ErrTokenExpired:
		respondError(w, http.StatusUnauthorized, err.Error())
	default:
		respondError(w, http.StatusInternalServerError, "Internal server error")
	}
}